requirements.txt
gradio
opencv-python
numpy
moviepy
pillow
scikit-image